+++ src/husaccttest/analyse/ArchitectureReconstructionTest_SRMA.java
@@ -22,8 +22,6 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.HashSet;
-import java.util.TreeMap;
-
 import org.apache.log4j.Logger;
 import org.apache.log4j.PropertyConfigurator;
 import org.junit.AfterClass;
@@ -61,7 +59,7 @@ public class ArchitectureReconstructionTest_SRMA {
 			mainController = controlService.getMainController();
 			while(isAnalysing){
 				try {
-					Thread.sleep((long)10);
+					Thread.sleep(10);
 				} catch (InterruptedException e) {}
 				isAnalysing = mainController.getStateController().isAnalysing();
 			}
@@ -78,7 +76,7 @@ public class ArchitectureReconstructionTest_SRMA {
 			mainController = controlService.getMainController();
 			while(isValidating){
 				try {
-					Thread.sleep((long)10);
+					Thread.sleep(10);
 				} catch (InterruptedException e) {}
 				isValidating = mainController.getStateController().isValidating();
 			}
