+++ src/husaccttest/analyse/CSharp_AccuracyTestDependencyDetection.java
@@ -18,7 +18,6 @@ import java.net.URL;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.TreeMap;
 
 import org.apache.log4j.Logger;
@@ -57,7 +56,7 @@ public class CSharp_AccuracyTestDependencyDetection {
 			mainController = controlService.getMainController();
 			while(isAnalysing){
 				try {
-					Thread.sleep((long)10);
+					Thread.sleep(10);
 				} catch (InterruptedException e) {}
 				isAnalysing = mainController.getStateController().isAnalysing();
 			}
