+++ src/husaccttest/analyse/ExportImportAnalysedModelTest.java
@@ -58,7 +58,7 @@ public class ExportImportAnalysedModelTest {
 			mainController = controlService.getMainController();
 			while(isAnalysing){
 				try {
-					Thread.sleep((long)10);
+					Thread.sleep(10);
 				} catch (InterruptedException e) {}
 				isAnalysing = mainController.getStateController().isAnalysing();
 			}
