+++ src/husaccttest/analyse/Java_AccuracyTestDependencyDetection.java
@@ -54,7 +54,7 @@ public class Java_AccuracyTestDependencyDetection {
 			mainController = controlService.getMainController();
 			while(isAnalysing){
 				try {
-					Thread.sleep((long)10);
+					Thread.sleep(10);
 				} catch (InterruptedException e) {}
 				isAnalysing = mainController.getStateController().isAnalysing();
 			}
@@ -65,7 +65,7 @@ public class Java_AccuracyTestDependencyDetection {
 			mainController = controlService.getMainController();
 			while(isValidating){
 				try {
-					Thread.sleep((long)10);
+					Thread.sleep(10);
 				} catch (InterruptedException e) {}
 				isValidating = mainController.getStateController().isValidating();
 			}
