+++ src/husaccttest/analyse/ReconstructAlgorithmTests.java
@@ -68,7 +68,7 @@ public class ReconstructAlgorithmTests {
 			mainController = controlService.getMainController();
 			while(isValidating){
 				try {
-					Thread.sleep((long)10);
+					Thread.sleep(10);
 				} catch (InterruptedException e) {}
 				isValidating = mainController.getStateController().isValidating();
 			}
