+++ src/husaccttest/define/DefineSarServicesTest_SRMA.java
@@ -59,7 +59,7 @@ public class DefineSarServicesTest_SRMA {
 			mainController = controlService.getMainController();
 			while(isAnalysing){
 				try {
-					Thread.sleep((long)10);
+					Thread.sleep(10);
 				} catch (InterruptedException e) {}
 				isAnalysing = mainController.getStateController().isAnalysing();
 			}
