+++ src/husaccttest/graphics/DrawingControllerTest.java
@@ -75,7 +75,7 @@ public class DrawingControllerTest {
 			mainController = controlService.getMainController();
 			while(isAnalysing){
 				try {
-					Thread.sleep((long)10);
+					Thread.sleep(10);
 				} catch (InterruptedException e) {}
 				isAnalysing = mainController.getStateController().isAnalysing();
 			}
